修改bug

+++ src/main/java/com/xiliulou/electricity/entity/NewUserActivity.java
@@ -42,11 +42,11 @@ public class NewUserActivity {
     /**
      * 活动开始时间
      */
-    private Integer beginTime;
+    private Long beginTime;
     /**
      * 活动结束时间
      */
-    private Integer endTime;
+    private Long endTime;
     /**
      * 奖励类型，1--次数，2--优惠券
      */
